session-5/cosmos/requirements.txt
azure-cosmos==4.7.0
python-dotenv==1.0.1